closes #4  - добавил подстановку данных
