+++ Recipista/app/src/main/java/ru/mail/hse/recipista/MainActivity.java
@@ -1,10 +1,8 @@
 package ru.mail.hse.recipista;
 
 import android.os.Bundle;
-import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
-import ru.mail.hse.recipista.client.MealClient;
 import ru.mail.hse.recipista.fragment.IngredientFragment;
 import ru.mail.hse.recipista.tag.Tags;
 
@@ -18,8 +16,5 @@ public class MainActivity extends AppCompatActivity {
                 .beginTransaction()
                 .add(R.id.main_layout, new IngredientFragment(), Tags.TAG_LIST)
                 .commit();
-
-        MealClient mealClient = new MealClient();
-        mealClient.getAllIngredients();
     }
 }
