+++ Recipista/app/src/main/java/ru/mail/hse/recipista/fragment/IngredientFragment.java
@@ -11,6 +11,7 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import ru.mail.hse.recipista.client.MealClient;
 import ru.mail.hse.recipista.repository.IngredientRepository;
 import ru.mail.hse.recipista.R;
 import ru.mail.hse.recipista.adapter.IngredientAdapter;
@@ -31,7 +32,7 @@ public class IngredientFragment extends Fragment {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         recycler = view.findViewById(R.id.recycler);
-        recycler.setAdapter(new IngredientAdapter(IngredientRepository.list()));
+        recycler.setAdapter(new IngredientAdapter(IngredientRepository.list(), new MealClient()));
         recycler.setLayoutManager(new LinearLayoutManager(requireContext()));
     }
